+++ lib/spring/test/acceptance_test.rb
@@ -576,6 +576,13 @@ module Spring
         assert_failure "bin/rails runner ''", stderr: "timed out"
       end
 
+      test "server boot timeout with env variable" do
+        app.env["SPRING_SERVER_COMMAND"] = "sleep 2"
+        app.env["SPRING_BOOT_TIMEOUT"] = "1"
+
+        assert_failure "bin/rails runner ''", stderr: "timed out"
+      end
+
       test "no warnings are shown for unsprung commands" do
         app.env["DISABLE_SPRING"] = "1"
         refute_output_includes "bin/rails runner ''", stderr: "WARN"